apple-app-store/requirements.txt
app-store-web-scraper>=0.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
